Separate LlamaKitMacros; Add @Tool Macro; Fix extern LLAMA settings
