+++ objc/include/LlamaObjC.h
@@ -2,12 +2,17 @@
 #define LlamaObjC_h
 
 #include <Foundation/Foundation.h>
-#include <llama.h>
-#include <LlamaModel.h>
-#include <LlamaContext.h>
-#include <LlamaSession.h>
 #include <GPTParams.h>
+#include <GPTSampler.h>
+#include <llama.h>
+#include <LlamaBatch.h>
+#include <LlamaContext.h>
+#include <LlamaModel.h>
+#include <LlamaSession.h>
 
-
+int LLAMA_BUILD_NUMBER = 0;
+char const * LLAMA_COMMIT = "unknown";
+char const * LLAMA_COMPILER = "unknown";
+char const * LLAMA_BUILD_TARGET = "unknown";
 
 #endif /* LlamaObjC_h */
